requirements.txt
flask==3.1.0
python-dotenv==1.0.1
psycopg2-binary==2.9.10
requests==2.32.3
google-auth==2.36.0
google-api-python-client==2.155.0
apscheduler==3.10.4
sqlalchemy==2.0.36
pytz==2024.2
gunicorn==21.2.0
python-telegram-bot>=21.0
